+++ dutch/app/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.3.48
+- added an points-over-time graph with distinct player colors that are also shown in the player fields and points table
+- automatically opens the points table after round one and the points graph after round three
+
 ## 1.3.47
 - added German